+++ FirebaseAdmin/FirebaseAdmin.Tests/Auth/FirebaseUserManagerTest.cs
@@ -32,6 +32,8 @@ namespace FirebaseAdmin.Auth.Tests
         private static readonly GoogleCredential MockCredential =
             GoogleCredential.FromAccessToken("test-token");
 
+        private static readonly string CreateUserResponse = @"{""localId"": ""user1""}";
+
         [Fact]
         public async Task GetUserById()
         {
@@ -297,10 +299,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task CreateUser()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             var uid = await userManager.CreateUserAsync(new UserRecordArgs());
@@ -313,10 +312,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task CreateUserWithArgs()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             var uid = await userManager.CreateUserAsync(new UserRecordArgs()
@@ -345,10 +341,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task CreateUserWithExplicitDefaults()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             var uid = await userManager.CreateUserAsync(new UserRecordArgs()
@@ -371,12 +364,147 @@ namespace FirebaseAdmin.Auth.Tests
         }
 
         [Fact]
+        public async Task CreateUserEmptyUid()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                Uid = string.Empty,
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserLongUid()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                Uid = new string('a', 129),
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserEmptyEmail()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                Email = string.Empty,
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserInvalidEmail()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                Email = "not-an-email",
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserEmptyPhoneNumber()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                PhoneNumber = string.Empty,
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserInvalidPhoneNumber()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                PhoneNumber = "1234567890",
+            };
+
+            var exception = await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Equal(
+                "Phone number must be a valid, E.164 compliant identifier starting with a '+' sign.",
+                exception.Message);
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserEmptyPhotoUrl()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                PhotoUrl = string.Empty,
+            };
+
+            await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserInvalidPhotoUrl()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                PhotoUrl = "not a url",
+            };
+
+            var exception = await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
+        public async Task CreateUserShortPassword()
+        {
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
+            var userManager = this.CreateFirebaseUserManager(handler);
+            var args = new UserRecordArgs()
+            {
+                Password = "only5",
+            };
+
+            var exception = await Assert.ThrowsAsync<ArgumentException>(
+                async () => await userManager.CreateUserAsync(args));
+            Assert.Null(handler.Request);
+        }
+
+        [Fact]
         public async Task UpdateUser()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var customClaims = new Dictionary<string, object>()
             {
@@ -402,10 +530,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task LargeClaimsUnderLimit()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var customClaims = new Dictionary<string, object>()
             {
@@ -422,10 +547,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task EmptyClaims()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             await userManager.UpdateUserAsync(new UserRecordArgs()
@@ -442,10 +564,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public async Task NullClaims()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             await userManager.UpdateUserAsync(new UserRecordArgs()
@@ -462,10 +581,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public void ReservedClaims()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
 
             foreach (var key in FirebaseTokenFactory.ReservedClaims)
@@ -487,10 +603,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public void UpdateUserNoUid()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var customClaims = new Dictionary<string, object>()
             {
@@ -507,10 +620,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public void UpdateUserInvalidUid()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var customClaims = new Dictionary<string, object>()
             {
@@ -528,10 +638,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public void EmptyNameClaims()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var emptyClaims = new Dictionary<string, object>()
             {
@@ -549,10 +656,7 @@ namespace FirebaseAdmin.Auth.Tests
         [Fact]
         public void LargeClaimsOverLimit()
         {
-            var handler = new MockMessageHandler()
-            {
-                Response = @"{""localId"": ""user1""}",
-            };
+            var handler = new MockMessageHandler() { Response = CreateUserResponse };
             var userManager = this.CreateFirebaseUserManager(handler);
             var largeClaims = new Dictionary<string, object>()
             {